requirements.txt
Django==1.7.3
Pillow==2.7.0
argparse==1.2.1
cffi==0.8.6
cryptography==0.7.2
django-appconf==0.6
django-autoslug==1.7.2
django-compressor==1.4
django-diplomat==0.2.0
django-haystack==2.3.1
djangorestframework==3.0.5
elasticsearch==1.4.0
enum34==1.0.4
paypalrestsdk==1.7.2
psycopg2==2.5.4
pyOpenSSL==0.14
pyasn1==0.1.7
pycountry==1.10
pycparser==2.10
requests==2.5.1
selenium==2.44.0
six==1.9.0
urllib3==1.10.1
wsgiref==0.1.2
